+++ CHANGELOG.md
@@ -40,6 +40,9 @@
 - **plotter:** Добавлена кнопка загрузки файлов выгрузок через интерфейс
   браузера ([#67](https://github.com/mixayloff-dimaaylov/gstma/pull/67))
 
+- **utils:** Добавлен скрипт для выгрузки данных в формате Parquet
+  ([#69](https://github.com/mixayloff-dimaaylov/gstma/pull/69))
+
 - **docs:** Добавлена справка по подключению _Grafana Unified Alerting_ к
   Telegram ([#58](https://github.com/mixayloff-dimaaylov/gstma/pull/58))